add a responsive navbar
