+++ app/components/common/footer/CurrentTime.js
@@ -2,7 +2,7 @@ export default function CurrentTime() {
   const currentYear = new Date().getFullYear();
 
   return (
-    <div className="m-auto text-center p-10">
+    <div className="m-auto text-center">
       <p className="italic"> © {currentYear} </p>
       <p className="italic">ESTRADAS - Rui Costa</p>
     </div>
