+++ app/components/common/navbar/navbar_small/NavBarLink.js
@@ -1,0 +1,26 @@
+import Link from "next/link";
+import NAVBAR_LINKS from "@/components/lists/navbar_links_map";
+
+export default function NavbarLink(props) {
+  const ARRAY = [];
+
+  for (let i = 0; i < props.array.length; i++) {
+    ARRAY.push(props.array[i]);
+  }
+
+  return (
+    <div className="bg-black h-max p-4">
+      {NAVBAR_LINKS["order"].map((item) => {
+        return (
+          <div className="font-thin">
+            <Link href={NAVBAR_LINKS[item].href}>
+              <p className="text-gray-300 hover:text-yellow-300 font-thin p-2">
+                {NAVBAR_LINKS[item].section}
+              </p>
+            </Link>
+          </div>
+        );
+      })}
+    </div>
+  );
+}
