+++ app/styles/globals.css
@@ -35,3 +35,15 @@ main {
 svg {
   color: rgb(186, 152, 13);
 }
+
+@media screen and (max-width: 770px) {
+  .menu-normal {
+    display: none;
+  }
+}
+
+@media screen and (min-width: 770px) {
+  .menu-hamburguer {
+    display: none;
+  }
+}
